Split out a lot of code to a PrivacyProtection class
