+++ src/bridge/IrcBridge.ts
@@ -43,6 +43,7 @@ import { getBridgeVersion } from "../util/PackageInfo";
 import { globalAgent as gAHTTP } from "http";
 import { globalAgent as gAHTTPS } from "https";
 import { RoomConfig } from "./RoomConfig";
+import { PrivacyProtection } from "../irc/PrivacyProtection";
 
 const log = getLogger("IrcBridge");
 const DEFAULT_PORT = 8090;
@@ -86,6 +87,7 @@ export class IrcBridge {
         channel: string;
         networkId: string;
     }>;
+    private privacyProtection: PrivacyProtection;
 
     constructor(public readonly config: BridgeConfig, private registration: AppServiceRegistration) {
         // TODO: Don't log this to stdout
@@ -192,7 +194,8 @@ export class IrcBridge {
             defaultTtlMs: 10 * 60 * 1000, // 10 mins
         });
         this.matrixHandler = new MatrixHandler(this, this.config.matrixHandler || {}, this.membershipQueue);
-        this.ircHandler = new IrcHandler(this, this.config.ircHandler, this.membershipQueue);
+        this.privacyProtection = new PrivacyProtection(this);
+        this.ircHandler = new IrcHandler(this, this.config.ircHandler, this.membershipQueue, this.privacyProtection);
 
         // By default the bridge will escape mxids, but the irc bridge isn't ready for this yet.
         MatrixUser.ESCAPE_DEFAULT = false;
@@ -806,18 +809,6 @@ export class IrcBridge {
         await promiseutil.allSettled(promises);
     }
 
-    /**
-     * Get a cached copy of all Matrix (not IRC) users in a room.
-     * @param roomId The Matrix room to inspect.
-     * @returns An array of Matrix userIDs.
-     */
-    public async getMatrixUsersForRoom(roomId: string) {
-        // TODO: This needs caching.
-        const bot = this.bridge.getBot();
-        const members = Object.keys(await this.bridge.getBot().getJoinedMembers(roomId));
-        return members.filter(m => !bot.isRemoteUser(m));
-    }
-
     public async sendMatrixAction(room: MatrixRoom, from: MatrixUser, action: MatrixAction): Promise<void> {
         const intent = this.bridge.getIntent(from.userId);
         const extraContent: Record<string, unknown> = {};
@@ -856,10 +847,10 @@ export class IrcBridge {
         throw Error("Unknown action: " + action.type);
     }
 
-    public async syncMembersInRoomToIrc(roomId: string, ircRoom: IrcRoom) {
+    public async syncMembersInRoomToIrc(req: BridgeRequest, roomId: string, ircRoom: IrcRoom, kickFailures = false) {
         const bot = this.getAppServiceBridge().getBot();
         const members = await bot.getJoinedMembers(roomId);
-        log.info(
+        req.log.info(
             `Syncing Matrix users to ${ircRoom.server.domain} ${ircRoom.channel} (${Object.keys(members).length})`
         );
         for (const [userId, {display_name}] of Object.entries(members)) {
@@ -876,7 +867,15 @@ export class IrcBridge {
                 await new Promise(r => setTimeout(r, ircRoom.server.getMemberListFloodDelayMs()));
             }
             catch (ex) {
-                log.warn(`Failed to sync ${userId} to IRC channel`);
+                if (!kickFailures) {
+                    req.log.warn(`Failed to sync ${userId} to IRC channel`);
+                    continue;
+                }
+                req.log.warn(`Failed to sync ${userId} to IRC channel, kicking from room.`);
+                this.membershipQueue.leave(
+                    roomId, userId, req, true,
+                    "Couldn't connect you to this channel. Please try again later.", this.appServiceUserId
+                );
             }
         }
     }
@@ -996,6 +995,7 @@ export class IrcBridge {
             if (!event.content || !event.content.membership) {
                 return BridgeRequestErr.ERR_NOT_MAPPED;
             }
+            this.privacyProtection.clearRoomFromCache(event.room_id);
             this.ircHandler.onMatrixMemberEvent({...event, state_key: event.state_key, content: {
                 membership: event.content.membership as MatrixMembership,
             }});
